requirements.txt
asgiref==3.7.2
attrs==23.1.0
black==23.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coverage==7.2.7
Django==4.2.2
djangorestframework==3.14.0
drf-spectacular==0.26.3
factory-boy==3.2.1
Faker==18.11.2
flake8==6.0.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.7.0
pluggy==1.2.0
pycodestyle==2.10.0
pyflakes==3.0.1
pytest==7.4.0
pytest-cov==4.1.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
referencing==0.29.1
rpds-py==0.8.8
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
